login fix

+++ lib/ui/login/model/user_details_response.dart
@@ -3,7 +3,7 @@ class UserDetailsResponse {
   String? userId;
   String? profilePicture;
   String? farmName;
-  int? pondSize;
+  double? pondSize;
   String? fiscalYear;
   bool? active;
   bool? approved;
